+++ src/components/MessageBox.js
@@ -1,6 +1,15 @@
-import React, {useState} from "react";
+import React, { useEffect, useState } from "react";
 import { Image, View } from "react-native";
 import { ListItem, Avatar } from "@rneui/base";
+import {
+  collection,
+  query,
+  where,
+  onSnapshot,
+  limit,
+  orderBy,
+} from "firebase/firestore";
+import { auth, database } from "../../config/firebase";
 
 import styles from "./utils/styles";
 import useIndicator from "../hooks/useIndicator";
@@ -9,7 +18,15 @@ import { Pressable, TouchableHighlight, TouchableOpacity } from "react-native";
 
 const profileImg = require("../assets/profile-picture.png");
 
-const MessageBox = ({ navigation, userStatus, friendStatus, setPrevModalVisible }) => {
+const MessageBox = ({
+  navigation,
+  userStatus,
+  friendStatus,
+  dataSnap,
+  isPrevModalVisible,
+  setPrevModalVisible,
+}) => {
+  const [msgPreview, setMsgPreview] = useState("Say Hi!");
   const userIndicator = useIndicator(userStatus);
   const friendIndicator = useIndicator(friendStatus);
 
@@ -21,23 +38,57 @@ const MessageBox = ({ navigation, userStatus, friendStatus, setPrevModalVisible 
     invisible: "#818181",
   };
 
+  useEffect(() => {
+    try {
+      const q = query(
+        collection(database, "messages"),
+        where("conversationID", "==", dataSnap?.convRef),
+        orderBy("createdAt", "desc"),
+        limit(1)
+      );
+
+      const unsubscribe = onSnapshot(q, (querySnapshot) => {
+        if (querySnapshot.empty) return;
+
+        querySnapshot.forEach((doc) => {
+          setMsgPreview(doc.data().text);
+        });
+      });
+
+      return () => unsubscribe();
+    } catch (err) {
+      console.log(err);
+    }
+  }, []);
+
   const [isModalVisible, setModalVisible] = useState(false);
   const [isIndividialModal, setIndividualModal] = useState(true);
 
   return (
     <TouchableHighlight
       onLongPress={() => setModalVisible(!isModalVisible)}
-      onPress={() => {navigation.navigate("Chat"); setPrevModalVisible(false);}}
+      onPress={() => {
+        navigation.navigate("Chat", {
+          convID: dataSnap?.convRef.id,
+          title: dataSnap?.title,
+        });
+        setPrevModalVisible(false);
+      }}
       delayLongPress={100}
       underlayColor={"rgba(217, 217, 217, 0.5)"}
     >
       <View>
-        <QuickAccessViewModal isModalVisible={isModalVisible} setModalVisible={setModalVisible} isIndividualModal={isIndividialModal} setIndividualModal={setIndividualModal}/>
-        
+        <QuickAccessViewModal
+          isModalVisible={isModalVisible}
+          setModalVisible={setModalVisible}
+          isIndividualModal={isIndividialModal}
+          setIndividualModal={setIndividualModal}
+        />
+
         <ListItem
-      //onPress={() => navigation.navigate("Chat")}
-      containerStyle={styles.listContainer}
-    >
+          //onPress={() => navigation.navigate("Chat")}
+          containerStyle={styles.listContainer}
+        >
           {/** TODO: optimize by converting into a component with small, medium, and large options */}
           <View>
             <Avatar size={61} rounded source={profileImg} />
@@ -47,13 +98,18 @@ const MessageBox = ({ navigation, userStatus, friendStatus, setPrevModalVisible 
           <ListItem.Content style={styles.wholePreview}>
             <ListItem.Content style={styles.row}>
               <ListItem.Title numberOfLines={1} style={styles.name}>
-                Cyril de Guzman
+                {dataSnap?.title}
               </ListItem.Title>
-              <ListItem.Subtitle style={styles.time}>11:10 PM</ListItem.Subtitle>
+              <ListItem.Subtitle style={styles.time}>
+                11:10 PM
+              </ListItem.Subtitle>
             </ListItem.Content>
             <ListItem.Content>
-              <ListItem.Subtitle style={styles.messagePreview} numberOfLines={1}>
-                Them: trying his best
+              <ListItem.Subtitle
+                style={styles.messagePreview}
+                numberOfLines={1}
+              >
+                {msgPreview}
               </ListItem.Subtitle>
               <ListItem.Subtitle
                 style={{
@@ -63,8 +119,8 @@ const MessageBox = ({ navigation, userStatus, friendStatus, setPrevModalVisible 
                 numberOfLines={1}
               >
                 They see you as{" "}
-                <Image source={userIndicator} style={styles.smallIndicator} /> until
-                tomorrow 8:30pm
+                <Image source={userIndicator} style={styles.smallIndicator} />{" "}
+                until tomorrow 8:30pm
               </ListItem.Subtitle>
             </ListItem.Content>
           </ListItem.Content>