+++ CHANGELOG.md
@@ -9,13 +9,12 @@ NOTE: all lowercase and no ending punctuation marks
 
 ## Next release v1.0
 
+- [CY] - [#22](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/22) app(feature): source message boxes from firestore
 - [PB] - [#32](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/32) app(screen): add bubble screen
 - [PB] - [#31](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/31) app(feature): settings
 - [PB] - [#23](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/23) app(feature): quick access view modal
-
 - [PB] - [#21](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/21) app(screen): set status for specific audience
 - [LE] - [#16](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/16) app(screen): new message screen
-
 - [CY] - [#18](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/18) app(screen): chat box screen
 - [PB] - [#24](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/24) app(feature): bubble see all modal screen
 - [PB] - [#25](https://github.com/cyril-deguzman/donotdisturb-instant-messenger/issues/25) app(feature): bubble screen